Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 260303/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -6,9 +6,9 @@ using namespace std;
 
 long long m;
 long long a, b;
-long long minAns = LLONG_MAX, maxAns = LLONG_MIN;
+int minAns = INT_MAX, maxAns = INT_MIN;
 
-long long SearchNum(long long num)
+int SearchNum(long long num)
 {
     int cnt = 0;
     long long left = 1, right = m;
